+++ tests/test_nested.R
@@ -5,3 +5,8 @@ form  <- log(gsp) ~ log(pc) + log(emp) + log(hwy) + log(water) + log(util) + une
 summary(plm(form, data = pProduc, model = "random", effect = "nested"))
 summary(plm(form, data = pProduc, model = "random", effect = "nested", random.method = "walhus"))
 summary(plm(form, data = pProduc, model = "random", effect = "nested", random.method = "amemiya"))
+
+pProduc_unbal <- pProduc[-c(2, 20:45, 75, 83:85, 500:510, 632:688, 700, 750), ]
+summary(plm(form, data = pProduc_unbal, model = "random", effect = "nested"))
+summary(plm(form, data = pProduc_unbal, model = "random", effect = "nested", random.method = "walhus"))
+summary(plm(form, data = pProduc_unbal, model = "random", effect = "nested", random.method = "amemiya"))